// types/enums.ts
export enum AppRoutesEnum {
  HOME = '/',
  EXPERIENCE = '/experience',
  PROJECTS = '/projects',
  SKILLS = '/skills',
  CONTACT = '/contact',
}

export enum AppRoutesLabelsEnum {
  HOME = 'Home',
  EXPERIENCE = 'Experience',
  PROJECTS = 'Projects',
  SKILLS = 'Skills',
  CONTACT = 'Contact',
}

export enum LanguageEnum {
  UKRAINE = 'ua',
  ENGLISH = 'en',
}